Redmine#7926: Empty network response is not always considered as an error
The motivation for the `ReceiveTransaction` change is the issues that can occur because some calls to it do not consider 0 as an error, and continue execution. It leads to misleading logs, not mentioning a networking issue, like in:
* The call at https://github.com/cfengine/core/blob/master/libcfnet/client_code.c#L591, where the unitialized `buf` will likely fail to be decrypted.

Or to actual issues like:
* The call at https://github.com/cfengine/core/blob/master/libcfnet/stat_cache.c#L237. If it returns 0 and the file is a symlink, the path of the link will be empty, and the destination link will be replaced by an empty link until next execution.

Currently the possible cases are:
* `recv` returns -1. `ReceiveTransaction` returns -1
* `recv` returns 0. This is not an error at the networking level, but it is at the CFEngine protocol level. `ReceiveTransaction` returns 0.
* `recv` returns >0, but with an unexpected (at the protocol level) length. `ReceiveTransaction` returns -1.

The rationale for the `ReceiveTransaction` change is to return -1 in all error cases, and to abstract the networking or protocol states, which is currently only partially done. The networking and protocol errors are all handled and logged in `ReceiveTransaction` (currently, the networking error logging relies on the caller).

Changelog: Redmine#7926: Fix handling of closed connections during transactions