Redmine#7912: Fix classes being set because of hash collision in the implementation.
The implementation was using a Red-Black Tree as a global class
table. Unfortunately the key was not the full class name, but just a
hash of that. As a result, when two classes had the same hash, only one
had to be set for both to appear as set.

The fix removes the Red-Black Tree from the global class table, and
replaces with a hash table, which handles collisions natively.

Changelog: Title
(cherry picked from commit 5a70c38474b6591993b40204f46e20ed341822d1)