redmine#7952: Fix use-after-free in ArrayMap and HashMap
When re-inserting a key-value pair into our two Map implementations
(ArrayMap and HashMap) then the old entry gets overwritten. However the
implementation only overwrote the value, not the key, since the two keys
compare as equal. Additionally it *freed the new key*.

Although the old and new keys have the same contents, they are different
pointers. And if the key is referenced within the value struct, then
both the value and key must be in sync. Else the value might reference a
freed key.

The fix was: in case of overwriting a pre-existing key-value pair, write
both the key and the value (despite the old and new keys comparing as
equal). Finally free the old key and the old value.

Changelog: Title