CFE-2161: Work around bugs in ps tools on Solaris.
It is not possible to get a correct process reading from one ps
command on Solaris. There are two tools, /bin/ps and /usr/ucb/ps (see
below), and both are wrong in their own particular way, and we need to
combine information from both of them to get a correct picture. Here
are the properties of the two ps tools:

- /bin/ps
  - Cannot report more than eighty 80 characters of the command
    line.
  - Everything else is correct AFAIK.
- /usr/ucb/ps
  - Reports full command line, beyond 80 characters.
  - Mangles size fields when the numbers get large, in a way that is
    impossible to recover from (I cannot even do it manually, the
    information is scrambled). This messes up the parsing, resulting
    in both error messages and wrong information.

What we do in this pull request is to use /bin/ps as the primary
source of information, but we use /usr/ucb/ps to augment that
information, providing the full 80+ command line.

We also change behavior slightly: Now we only match the promiser
against the command line of the process, not the whole ps line.

- Pros:
  - Will match the full 80+ cmd line, regardless of the line that
    either ps tool outputs.
  - It's more intuitive that the promiser is matched against the
    command line of the process.
  - The ps line is never directly exposed anywhere, reducing platform
    complexity, making it more uniform.
- Cons:
  - If anyone is explicitly using the promise to match against other
    fields than the command line, this will stop working.

Note that this behavior change is made globally for all platforms.

This also requires changes to some tests:

- It is no longer possible to match for a PID in findprocesses() and
  processexists(). These functions are new, so that's no problem, if
  needed this functionality can be added later.
- The process promise can match PIDs, but now needs to do it using the
  dedicated `pid` attribute. OTOH, see how much cleaner the regex
  became because of it!

Changelog: Improve robustness of process table parsing on Solaris.

Changelog: Behavior change: The promise string of a `processes`
promise now matches just the command line of each process instead of
the line that is output by ps. This was done to reduce fragmentation
between platforms, since ps is a very nonstandardized tool.