Redmine #7961: Remove auto-loading of "inputs" mentioned in def.json.
This was done in order to solve the problem that def.json is applied
to all policy files, including failsafe.cf. If there is a syntax error
in any of those files, all policy files will stop working and the host
is unable to recover.

Now one has to use the "inputs" section by referring to it as the
$(def.augments_inputs) variable, which it will populate.

Changelog: Fix bug which could render host unable to recover from a
syntax error, even if failsafe.cf was utilized. This could happen if
the file containing the syntax error was specified in the def.json
special file.

Changelog: Change: Policy files specified in the "inputs" section of
def.json will no longer be auto-loaded. One has to refer to the
$(def.augments_inputs) variable in the policy (the standard
masterfiles policies include this by default). This only affects
installations which are not based on the standard masterfiles, and
which are using the "inputs" field inside def.json.