Global classes that caused agent to abort were previously not printed, the regex that caused them to abort was. This was not useful and inconsistent with local classes (the class itself is printed). Semantics of GetAgentAbortingContext() was changed, but it was only used in this one place.
ChangeLog: Classes matching agent control's abortclasses are
now printed before exit, even if they are defined in common bundles.
Previously the regex (in abortclasses) that matched the class was
printed if the class was defined in a common bundle, but the class
itself was printed if it was defined in an agent bundle. With this
change, the defined class that caused the abort is always printed.