CFE-2183: Figure out version instead of hardcoding.
This way we can build all versions we want automatically, without
having to hardcode the version number in the source, using the
EXPLICIT_VERSION variable. If the variable is not set, and git is
available, it will try to figure it out from git tags.

Note that all of this is done in the autogen stage, because git will
generally not be available at the configure stage. This is also
required in order for tarballs to come with the correct version
embedded in them, because they won't have git available to figure this
out.
(cherry picked from commit fb36b03a6b7979092a9e515fa9f26b3f87bcf758)

Conflicts:

	configure.ac
	tests/acceptance/00_basics/macros/if_triple.cf