Cleanup signal handling for agents and enable them earlier.
The motivation for this is that during policy loading, certain
functions may be evaluated that open pipes, and if these pipes at any
point become disconnected, we get an EPIPE signal. This is ignored
once we get to the evaluation stage, but during loading the signal
handlers have not been enabled yet.

This was originally seen during introduction of the mapdata()
function, which opens a pipe to an external process. In the test for
this function, `cat` was used to output a prerendered response, but
`cat` was also expected to consume the data input. Depending on
timing, if the data was written to `cat` before it terminated,
everything would be fine, but if `cat` terminated before data could be
written (because it was catting a file, not from stdin), cf-promises
would get an EPIPE signal and terminate abnormally.

Upon inspection of our signal handling, it appears this can happen in
cf-agent as well, and is as such "a ticking time bomb". Therefore this
commit cleans up the signal handling a bit, consolidates it where
possible, and moves all signal handler initialization to the beginning
of each agent main().

Daemons have other signal handler requirements and were left alone for
now.