Add "true" and "false" special logic expressions.
Add "true" as a special expression that is always EVAL_EXPRESSION_TRUE.
Add "false" as a special expression that is always
EVAL_EXPRESSION_FALSE. This only happens in context expression
evaluations, testing if a class is defined.

Changelog: behaviour change: "true" is always defined and "false" is never defined in a context expression.