Reverse order of binary lookup in testall.
It's actually the last path that ends up being used, not the first,
and it is this order of priority we want. This was motivated by trying
to run acceptance tests in a directory not named core, but core was
present next to it.
(cherry picked from commit de4edbc16b0dd23599d88b8d122e59e7197437ce)

Conflicts:

	tests/acceptance/testall