CFE-2283: Add email filtering feature for cf-execd.
Also add some tests for cf-execd, where we mock one function which is
the connection to the SMTP server. Instead we connect to a local pipe
and feed it data.

Changelog: New feature: Emails sent by cf-execd can be filtered to get
rid of emails for unwanted log messages. The attributes
`mailfilter_include` and `mailfilter_exclude` in `body executor
control` control what to include. See documentation for `cf-execd` for
more information.