Fix "No such file or directory" LMDB error on heavily loaded hosts.
There is a race condition in LMDB that will fail to open the database
environment if another process is opening it at the exact same
time. This condition is signaled by returning ENOENT, which we should
never get otherwise. This can lead to error messages on a heavily
loaded machine, so try to open it again after allowing other threads
to finish their opening process.

Ref: CFE-2300

Changelog: Title