CFE-2263 Fix select_end EOF matching.
Make sure that if select_end is not matched in file, EOF is used as
end of region only if 'select_end_match_eof' is set to true.
The default value is false, so EOF is not matched as end of region
by default.

Changelog: EOF is  matched as an end of the region in edit_line
promises only if 'select_end_match_eof' parameter is true. (CFE-2263)
(cherry picked from commit 673f3a8a1c84e9852f0dde9fcdc47ebae41b83f0)