CFE-2263: Unify delete_lines and insert_lines behavior.
If select_end is not matched in insert_lines promise EOF
won't be implicitly used as the end of the region unless
'select_end_match_eof' is true.

Changelog: insert_lines is no longer implicitly matching EOF as
end of the region if 'select_end' pattern is not matched (CFE-2263).