IntFromString() and DoubleFromString() previously used a string 'remainder' and called IsSpace(remainder).
In a previous commit, remainder was converted into a char
and isspace(remainder) was used instead. However, this
introduced a subtle issue because IsSpace(remainder)
returns true if remainder is empty (or just space).

This commit amends this by checking both for space and empty
character. Also adding some tests to try avoiding this in the
future.