Check for struct stat.st_blocks properly.
One statement inside CopyRegularFileDisk() was ST_BLKSIZE(dstat), with
dstat nowhere defined. Investigating this uncovered that this macro is
never properly defined, since it was based on HAVE_ST_BLOCKS macro, and
the relevant autoconf initialisation routine was never called.

So the appropriate check for stat.st_blocks was added in configure.ac,
and then compilation immediately failed (because of dstat being
undefined). This was fixed and block counting modernised.

Note that AC_STRUCT_ST_BLOCKS macro could not be used in configure.ac
because it required a special libcompat/fileblocks.c to be
present. Anyway the recommended way is to use AC_CHECK_MEMBERS() like in
this patch.