ENT-2769: Speed up local and remote file copying and fix spurious errors.
Changelog: Title

When writing to disk files, zeroes are automatically being detected and
sparse files are being created to save disk space. This was happening
even for small "holes" of a couple of zeroes, and the system call
overhead was high.

It was fixed by writing "holes", i.e. making the file sparse, only if
the holes of zeroes were longer than the block size.

Also fixes a seemingly spurious failure to write files (ftruncate() was
returning EPERM). Reason is that on files that ended with zeroes, we
used to call ftruncate() with an offset past the end of the file. This
is not guaranteed functionality under POSIX, and indeed it was failing
on Linux under non-native filesystems (vfat, vboxfs).

This was fixed by write()ing one byte and then ftruncate()ing it back.