CFE-2397: Avoid JSON incompatible code path in getindices().
If we have two indexed CFEngine variables:

vars:
  "a[x]"    string => "one";
  "a[x][y]" string => "two";

this structure is impossible to turn into JSON because if we do, we
get:

{
  "a": {
    "x": "one",
    "x": {
      "y": "two"
    }
  }
}

which contains two identical "x" keys, and is therefore invalid JSON.

In this case the old code would crash in some cases, but it's order
dependent, so it didn't always happen.

It's not necessary to use JSON for getindices(), since it returns a
list, but this was nevertheless the path taken in the code. So fix it
by avoiding this code path and taking the legacy path which handles
this case. Some extra care is needed because there are several ways in
which the input can be JSON, either by being specified inline, or by
referring to a variable that holds JSON, and in this case we need to
take the JSON path.

Changelog: Fix occasional segfault when running getindices() on a
variable that has indices of multiple depths (e.g. both "a[x]" and
"a[x][y]").
Signed-off-by: Kristian Amlie <kristian.amlie@cfengine.com>

(cherry picked from commit b76b4cd34eb3d2f5bf179ba7b6565d4525d2e6fc)