Adapt autoconf code to autoconf 2.63.
Autoconf 2.63 is lacking both m4_ifblank and m4_esyscmd_s, so work
around the former with m4_ifval (which has reversed semantics) and the
latter with m4_normalize.

Also bump the version requirement while we're at it. Although I don't
know if the new macros were introduced in 2.63, we at least know they
are present there. Bump to 2.63 because:

* 2.64 has been (secretly) required for a long time, since the macros
  I replaced in this commit were only available in that version and
  later. I doubt anyone is using anything older.

* 2.63 is already eight years old.

* At least 2.63 is tested, we don't know about the older ones.

* It doesn't affect tarballs.