Fix bug occuring when first user/group in the promised list was invalid.
Previously it would just skip the promise, now it searches for the first
valid user/group listed in the slist.

Also print better messages (CFE-2432) and when it makes sense
(previously a message for invalid user/group was printed only if the
promise had a single user/group listed in the slist).

Changelog: Fix bug in files promise when multiple owners are promised
but first one doesn't exist, and improve logging (CFE-2432).