CFE-2455: Fix ghost cf-serverd process remnants started in the chroot
This execresult() is being executed 4-5 because of the different
evaluation passes. The first times, the .cf file doesn't even exist. As
a result, failsafe.cf is created and executed, which launches
cf-serverd.

The fix here is necessary after the CFE-2162 changes were introduced, to
avoid running the command before populating the variable.