Fix spurious failures of short_expireafter timed test on Solaris 9.
It used to be that at second 0 a long-running promise is started, and at
second 65 the promise is re-run after killing the old one, because of
expireafter of 1 minute.

Spuriously though it happens that the promise has not expired, thus it
is not killed from the second run. This happens because the test
executes the promise as a sub-agent under a shell with ampersand (&) at
the end. If this happens at time 0, it wouldn't be surprising for the
sub-agent to need more than 5s to lock the promise, thuse failing to
expire at time 65.

Therefore we give a little more time, by changing the time of the
sub-agent re-execution after 75 seconds.