CFE-1563: Fix logic to detect when running under a Xen Hypervisor
On Linux the Xen_Hv_Check() function can never be called because the stat on
/proc/self/status will always return true which results in the OpenVZ_Detect()
function always being called. Move the call to Xen_Hv_Check() to happen if the
stat on /proc/xen/capabilities returns false since this is an indication that
the host is acting as dom0.

Change the Xen_Hv_Check() function to scan the CPUID range from
0x40000000 to 0x40010000 looking for Xen leaves.

Return false if the class redhat_4 or centos_4 is defined to avoid a segfault
on these platforms.

Take a look at this for some additional details about the fix:
https://github.com/mirage/xen/blob/master/tools/misc/xen-detect.c#L61

Changelog: Title