Unpluck unused bodies and bundles from test framework
While working on CFE-2519 the presense of manage_variable_values_ini in
the plucked library was being used instead of the MPF and no error was
thrown since the body was found complicating test development.

None of the bodies or bundles removed from the plucked library were used
in any tests.