CFE-2524: Avoid copying variable table when expanding a promise
The idea is to stop copying variable tables, which appears to be
one of the most heavy operation the agent does, but to always
lookup the variables in the global table.

The main issue with this change is to manage ambiguities
in the use of the "this" scope. It is used both for special
variables at promise level and as an alias for variables in the
current bundle.

The general idea is to modify VariableResolve to try to resolve
unknown "this"-scoped variables as variables in the current bundle.

The is a second change to force "this"-scopes variables
passed as parameters to a bundle to be qualified explicitly in
the calling bundle.

A non-compatible change could be to stop supporting "this" scope
for non-special bundle, which would allow a clean separation
and avoid the workarounds.

Changelog: Title