Add Comment to only call async-signal-safe functions in forked child.
Because these are generic functions that being called from
multi-threaded programs (cf-serverd, cf-execd etc), we are not allowed
to call any function after fork() and before exec(), because for example
the mutexes might be in random state. Note that this is true although
the fork()ed child is always a single-threaded process. See man 3p fork.