Empty signal mask after fork() but before exec().
Otherwise the child of cf-execd (which is cf-agent which possibly
executes another cf-execd) becomes unkillable!

Happened during ENT-3147: since cf-execd child-thread was set to ignore
all signals and because the thread's signal mask is inherited to the
fork()ed and exec()ed process, the `exec_command` was ignoring TERM etc
signals.