ENT-3267: fix segfault of RHEL4-built x86_64 package, on non-RHEL4 systems.
Xen_Cpuid() was causing stack corruption because the asm instruction
"push %rbx" was modifying the red zone. On x86_64 SystemV ABI, leaf
functions are allowed to use -128(%rsp) ... (%rsp) without updating the
stack pointer %rsp. And the RHEL4-specific GCC was generating code that
saved %rbx and %rbp in this red zone, at the beginning of Xen_Cpuid(),
and restored them in the end. Then the value of %rbp got corrupted (with
the value of %rbx because of the "push %rbx" instruction). This happened
for all un-optimised builds only; on optimised builds, Xen_Cpuid() was
inlined, and the red-zone rules would not apply since the containing
function was not a leaf function.

Read more at this excellent post:

https://stackoverflow.com/questions/34520013/using-base-pointer-register-in-c-inline-asm/34522750#34522750

To fix I just removed the push/pop instructions from x86_64
assembly. Preserving %ebx is only needed for i386 PIC code, no such
thing is needed on x86_64.