CFE-1960: fix storage mount promise when existing mountpoint has a similar path
The MatchFSInFstab will parse the each line of the fstab looking for the
existence of a mountpoint as a substring. In certain cases the call to strstr
would incorrectly match if a nfs volume with a similar mount point was already
mounted. Instead of using strstr, tokenize the line and look at each token for
an exact match.

ChangeLog: Title