sha-deletion.sh wrapper script for record cleanup
For bulk removals of host records this script is incomparably easier
to use than the UI or the REST API.  For single host removals it is
still slightly easier than the REST API and there is lot less to remember.

I've been using this for many months; there are really no caveats.  Input is
fully validated before use.  (It's not checked that a given SHA actually *exists*
on the hub before attempting to remove it; only that is validly formatted as a SHA.
However, this matches the behavior of the UI for hostkey removals, so not an issue.)