Add permission statement to cfengine-code-style.el
It is the convention to have such a header in every Emacs lisp
library.  One benefit is that it helps tools that extract Emacs
packages from repositories that contain much more than just the
Emacs lisp.  Without this change the tools used to maintain the
Emacsmirror does not see the license because it only extracts
the library itself from the repository and discards all other
history.