CFE-1782: Fix bug preventing permission changes on Unix sockets.
Making a test for a Unix socket is considerably more effort than for a
FIFO (needs a server to run), so I'm not including that. Most of the
conditions for a socket should be the same as for a FIFO anyway.

Changelog: Title
Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>
(cherry picked from commit d9deff6415b06a1cfec24aee75ac92e0cd299ee0)