ENT-3483: Do not tag large volatile variables for inventoy
sys.interfaces_data, sys.inet and sys.inet6 are commonly larger than the
maximum data size allowed to be collected by cf-hub. Data larger than 1k
is truncated. Instead of reporting truncated data this change stops
tagging the variable so that it will not be collected to the Enterprise
hub and will not be available in Mission Portal.

Changelog: Commit

ENT-3483: Do not tag sys.inet and sys.inet6 for inventory