Change how maparray() and mapdata() tests work
Documentation for both maparray() and mapdata() says that "The
order of the keys is not guaranteed.". While maparray() returns
an slist which can be easily sorted with sort(), mapdata()
returns data, more specifically a JSON array of various things
and cannot be sorted with sort().

So when testing mapdata() a file with expected results cannot be
used. Instead, we need to check that everything that's expected
to appear in the results is actually there. Which is a complex
thing to do and requires many individual checks.