CFE-2748: Issue a warning on ignored locking attributes
According to the recently updated documentation on promise
locking, there are several promise types that don't support any
form of promise locking:

  * access
  * classes
  * defaults
  * meta
  * roles
  * vars

When a policy writer uses "ifelapsed" or "expireafter" attributes
(via an action body) on these promise types which don't support
them, a warning should be issued rather than the setting being
silently ignored.

Codewise, it would be great if every promise evaluation called
GetTransactionConstraints(), that way we could only have the
check and warning in one place. Unfortunately, that's not the
case, VerifyVarPromise() is not calling that particular function
and neither is handling of 'access' and 'roles' in cf-serverd.

Changelog: Title