Mark shuffle() tests as expected to fail on (centos_4|debian_4).64bit
For some reason, shufle() produces different results on 64bit
RHEL/CentOS 4 and Debian 4 than everywhere else. Even when the
same seed and the same hashing algorithm are used. This is
probably due to some differences in glibc or maybe sizes of some
types. Anyway, it's only these two platforms differing from all
the others and results *are* still consistent in multiple runs on
a single platform. So it seems like this can be safely ignored.