fix detection of libacl headers
Commit 41c0afd ("Improve readability") moved the error handling for a
missing libacl header into AC_CHECK_HEADER's action-if-not-found.  This
was a mistake, though, as this action is executed for each header not
found, which was not the intent of the original check.

For libacl, it is enough for any of the header files [acl.h sys/acl.h
acl/libacl.h] to be present.