Create cf-diag.sh
This is a standalone diagnostic shell script that should run properly on ANY system that can run CFEngine.
It is intended for operator use in quickly diagnosing a system.  However, it uses "test anything protocol" so can
also be used from any "harness" designed for that.  (https://testanything.org/)

One test ("cf-agent ran in last 15 minutes") may falsely fail on systems without GNU coreutils (I depend on "stat"
for that test), but otherwise I would expect this script to work on a wide variety of OSes.