CFE-2125: Do not iterate over JSON objects' properties in mustache
Only lists are supposed to be iterated over. Otherwise duplicates
appear when using JSON with objects that have multiple
properties. See CFE-2125 for details.

However, our '{{@}}' extension to the mustache specification
requires JSON objects' properties to be iterated over because
it's the names of these properties (keys) it should be
substituted with.

Changelog: title