libpromise: qdbm: fix DBPrivOpenCursor function
Since commit 7b05ffecf (Add support for new packages promises in dbm.,
2015-05-20), the function `DBPrivOpenCursor` was completely broken.
Ignoring the fact that the function is recursively calling itself until
the stack overflows and cfengine segfaults, the only thing it would do
if it succeeded in opening a cursor would be to delete the complete
database. This code cannot have ever been executed by anyone, as there
is clearly no constellation where it could possibly work.

I can only assume that this was an unintentional change to
`DBPrivOpenCursor`. Thus, this commit reverts all changes made in that
commit to `DBPrivOpenCursor`.