CFE-741 Do not error when file does not exist with isexecutable()
Changelog: Title

Since isexecutable() is only checking to see if a file is executable we do not
need to have an error message that is always emitted if the file does not exist.
The function will return non-true and the class will not be set. If the file
does not exist, it's not executable.

(cherry picked from commit 0df7a2a3fa18e710a24c1c040d5bb2f43e73aa1a)