CFE-2793 Identify bundle name when triggering bundle abortion
Changelog: Title

This change is intended to make it easier to identify what bundle was aborted
when `abortbundleclasses` are triggered.

Now when a bundle is aborted the log message says Bundle x aborted on defined
class y. Previously when a bundle was aborted the log message only indicated the
class that triggered the abortion.