ENT-3756: cf-agent intermittently hangs on windows
Refactored use of atexit() to use of a custom cleanup mechanism. On Windows atexit() functions
are called during the unloading process and we have some atexit() functions which rely on
DLLs so this will not work reliably. Refactored so all platforms use RegisterCleanupFunction()
and DoCleanupAndExit() instead of atexit() and exit().