Add a test that remote variable injection using module works
As described in CFE-1915, this is one of the possible work
arounds for cases that absolutely need and cannot avoid remote
variable injection (defining a variable in a remote
bundle). However, it should be clear that it is not a good
practice.

(cherry picked from commit ed705b99bb0ddc223e7f5a686f0784827d0f8de1)