Give processes time to terminate properly in the init script
Our processes try to clean after themselves if terminated
correctly. But it takes some time so sending the TERM signal
twice in a row with no time in between doesn't make sense. There
has to be some time after the first signal for the process to do
its cleanup.

Ticket: CFE-2835