Add a new 'off_the_record' connection flag and use it in cf-net
Sometimes we don't want the connection to be recorded in the
lastseen DB. There might be multiple reasons for this like
peformance or avoiding ruining the lastseen DB records when
debugging network issues. Both of these cases apply to the
'cf-net' tool which should thus use this new flag.

Ticket: CFE-2846