Fix the argument for the '-b' option of cf-runagent to really be optional
The '--help' output says it is an optional argument, and the code
says 'optional_argument' as well. But for this to work with
'getopt_long()' two colons have to be used after the short option
character.

(cherry picked from commit 0b4e8adfdf8c9917d99377547b2ef132441d0d5b)