CFE-2880: Add support for setting package module interpreter
Adding hashbangs to package modules implemented as scripts is
suboptimal and potentially quite complicated (python2
vs. python3, etc.). This can be much nicely determined in the
policy and thus we should provide a way to set the interpreter
for a package module. This commits adds a new attribute
'interpreter' to the 'package_module' body.

Changelog: Title