CFE-2103: Add support for overriding the package module's path
Sometimes it is useful to define a new package module that is
just a variation of a different one. Also package modules can be
scripts with various names, e.g. ending with file extension like
".py" or ".sh", but package module names are quite restrictive.

Enforcing the package module implementations to be executables or
scripts under the /var/cfengine/modules/packages/ with the same
name as the package modules (body package_module NAME) is thus
quite limitting and we need to provide an easy way to override
the package module path.

Changelog: Title