Changed functions in verify_files to use Attributes *
Attributes struct is 2304 bytes and shouldn't be copied.
There were a few places where the local copy is edited,
so I added a copy inside the function instead, with
TODO comment to try to get rid of this later.

In these cases it's hard to determine what is correct / intended.
(Did the author intend to edit the local copy or not?)

Reported by LGTM:
https://lgtm.com/rules/2163210742/

Ticket: CFE-2886
Changelog: None
Signed-off-by: Ole Herman Schumacher Elgesem <ole@northern.tech>