Fixed small memory leaks of environment variable strings
Discovered using valgrind.

I had to introduce an alternative to putenv(),
putenv_static() which stores it's argument in
a static sequence so it can be freed later.
I wanted to use setenv(), but it is marked as
deprecated, and not portable.

This change means that env strings are always
reachable. (Until they are destroyed by
putenv_static_destroy())

Changelog: Title
Ticket: CFE-2918
Signed-off-by: Ole Herman Schumacher Elgesem <ole@northern.tech>