Fixed valgrind warning about lmdb writing uninitialized memory
This was interesting. The lockdata struct had uninitialized
padding which was written to disk by LMDB. In C, even if you
initialize a struct by specifying the values of all members,
the padding between them will not be initialized. This caused
the long standing LMDB warning about writing uninitialized
memory. It is most likely completely harmless, but very
good to get rid of this noise in valgrind output.

Ticket: CFE-2591
Changelog: Title
Signed-off-by: Ole Herman Schumacher Elgesem <ole@northern.tech>