Added a new binary: cf-check
Corrupt local databases (LMDB) continues to be a problem.
cf-check will be used to diagnose and remediate problems
with corrupt databases. It is a standalone binary, which
doesn't evaluate policy or use the local databases, thus
it can be used in situations where the other binaries
like cf-agent would hang.

cf-check replaces our lmdb database dumper, lmdump.
`cf-check lmdump` or symlinking / renaming it to lmdump
will make cf-check have the exact same behavior as lmdump.

cf-check will include much more functionality in the future
and some of the code will be added to other binaries,
for example to do health checks of databases on startup.

Ticket: ENT-4064
Changelog: Commit
Signed-off-by: Ole Herman Schumacher Elgesem <ole@northern.tech>