Add parsed policy files hashes/digests to Policy
We are computing hashes/digests of policy files when loading them
to detect duplicate policy files etc. However, this information
is useful even later, e.g. for logging where it allows us to tie
the promise outcome with a particular version of the policy file
the promise appeared in.

Ticket: ENT-3996