Allow up to 5 file descriptors per connection in cf-serverd
2 FDs availalbe for a connection/thread being enough is really
hard to imagine. There's the socket so that leaves 1 FD for
anything else. Then every connection is being written into the
lastseen.lmdb so that leaves 0 FDs for anything else. This
particular FD gets released again, but it requires nothing else
being opened by the thread in that moment.

Looks like OpenSSL 1.1.1 in some cases opens /dev/random and/or
/dev/urandom for the connection and then everything blows up.

But in general, we should allow more FDs per connection/thread
because they serve files etc. Let's try 5 for now, maybe we will
need even more in the future.

The original value (2 * MAXCONNECTIONS + 10) comes from the
ticket CFE-2575 where the problem was that the max number of FDs
in cf-serverd didn't reflect 'maxconnections' at all and thus the
default (which is 1024 on most systems) was insufficient for
systems handling many connections in parallel. Why the resulting
calculation was made to make the resulting numbers of max FDs
basically as small as possible is a mystery.

Changelog: Title