Refactor how the minimum TLS version is determined
There are multiple inputs:

* The version of the OpenSSL library we are compiling and linking
  with determines the set of available TLS versions (TLS 1.0-1.3
  in our builds).

* We define the minimum required TLS version (TLS 1.0 now).

* The minimum required version can be specified in policy.

And multiple rules we have:

1) Always use at least the minimum required version of TLS (1.0
   now, but that may/should change in the future).

2) If the minimum version requested by the policy is higher than
   the highest possible supported by OpenSSL, we use the highest
   possible version (and thus the best we can).

3) If an unrecognized minimum version of TLS is requested by the
   policy, we default to the minimum required (by us) version.

Both cases 2) and 3) produce a warning. The reasoning behind
those cases is that we do our best while keeping things work. If
we applied very strict rules and for example terminated instead
of using a lower minimum version than requested by the policy we
could easily lock out the clients. Fallback mechanisms are needed
to make sure the whole infrustructure has a chance to heal itself
once a valid and supported value is provided.